fix to support github style short dependencies
When a dependency of the form "username/repo#commit" is used as a
dependency of a module, npm expands this to
"git+https://github.com/username/repo#commit" in the dependencies list
in package.json but the `_from` value in that dependencies package.json
just has "git://github.com" causing the two values to mismatch when in
fact they are the same.

This commit simply fixes up the git+https:// url to just be git:// so
the two urls match.