Update node package semver
There is [a CVE](https://nodesecurity.io/advisories/31) for semver 4.2.0
(really it is for < 4.3.3) and so this PR updates to version 4.3.6.

Also updated the README to include the commands to run tests and checks
for CVEs in dependencies. TravisCI will also run package scans during
the testing phase on each build.